feat(aci): Detector sidebar setup
Further setup on the detector sidebar. Attempts to share some code between the details sidebar and the builder.
