+++ static/app/views/detectors/components/detailsPanel.tsx
@@ -1,3 +1,4 @@
+import {Fragment} from 'react';
 import styled from '@emotion/styled';
 
 import {Flex} from 'sentry/components/core/layout';
@@ -21,10 +22,14 @@ function SnubaQueryDetails({dataSource}: {dataSource: SnubaQueryDataSource}) {
       <Flex direction="column" gap={space(0.5)}>
         <Heading>{t('Query:')}</Heading>
         <Query>
-          <Label>{t('visualize:')}</Label>{' '}
+          <Label>{t('visualize:')}</Label>
           <Value>{dataSource.queryObj.snubaQuery.aggregate}</Value>
-          <Label>{t('where:')}</Label>{' '}
-          <Value>{dataSource.queryObj.snubaQuery.query}</Value>
+          {dataSource.queryObj.snubaQuery.query && (
+            <Fragment>
+              <Label>{t('where:')}</Label>
+              <Value>{dataSource.queryObj.snubaQuery.query}</Value>
+            </Fragment>
+          )}
         </Query>
       </Flex>
       <Flex gap={space(0.5)}>
@@ -63,9 +68,8 @@ const Heading = styled('h4')`
 const Query = styled('dl')`
   display: grid;
   grid-template-columns: auto auto;
-  grid-template-rows: repeat(3, 1fr);
   width: fit-content;
-  gap: ${space(0.5)} ${space(1)};
+  gap: ${space(0.25)} ${space(0.5)};
   margin: 0;
 `;
 
@@ -76,7 +80,7 @@ const Label = styled('dt')`
 `;
 
 const Value = styled('dl')`
-  ${p => p.theme.overflowEllipsis};
+  word-break: break-all;
   margin: 0;
 `;
