+++ static/app/views/detectors/detail.spec.tsx
@@ -3,6 +3,7 @@ import {DetectorFixture, SnubaQueryDataSourceFixture} from 'sentry-fixture/detec
 import {OrganizationFixture} from 'sentry-fixture/organization';
 import {ProjectFixture} from 'sentry-fixture/project';
 import {TeamFixture} from 'sentry-fixture/team';
+import {UserFixture} from 'sentry-fixture/user';
 
 import {render, screen} from 'sentry-test/reactTestingLibrary';
 
@@ -52,6 +53,10 @@ describe('DetectorDetails', function () {
         AutomationFixture({id: '2', name: 'Automation 2'}),
       ],
       match: [MockApiClient.matchQuery({id: ['1', '2']})],
+    });
+    MockApiClient.addMockResponse({
+      url: `/organizations/${organization.slug}/users/1/`,
+      body: UserFixture(),
     });
   });
